Move TracePatch to own file
